refactor: change package names
- Updated package names from `com.android.sample` to `com.arygm.quickfix`
- Ensured all imports and references are updated accordingly
