+++ app/src/main/AndroidManifest.xml
@@ -13,12 +13,12 @@
         android:theme="@style/Theme.SampleApp"
         tools:targetApi="31">
         <activity
-            android:name=".SecondActivity"
+            android:name="com.arygm.quickfix.SecondActivity"
             android:exported="false"
             android:label="@string/title_activity_second"
             android:theme="@style/Theme.SampleApp" />
         <activity
-            android:name=".MainActivity"
+            android:name="com.arygm.quickfix.MainActivity"
             android:exported="true"
             android:label="@string/title_activity_main"
             android:theme="@style/Theme.SampleApp">
