+++ app/src/main/java/com/arygm/quickfix/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.arygm.quickfix.ui.theme
 
 import androidx.compose.ui.graphics.Color
