+++ app/src/main/java/com/arygm/quickfix/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.arygm.quickfix.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
