+++ app/src/test/java/com/arygm/quickfix/PointTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.arygm.quickfix
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
